requirements.txt
streamlit ==0.86.0
aiortc==1.2.1
av==8.0.3
mediapipe ==0.8.6.2
opencv-python ==4.5.3.56
opencv-python-headless==4.5.3.56
pandas >=0.25.3
numpy   >= 1.19.2
Pillow ==8.1.0
streamlit_webrtc==0.26.0
typing_extensions==3.10.0.0